Updating to latest FlutterFlow output.

+++ lib/pages/home_page_m_s_f_s/home_page_m_s_f_s_widget.dart
@@ -24133,16 +24133,15 @@ class _HomePageMSFSWidgetState extends State<HomePageMSFSWidget> {
                                                                                                               _model.selectedRunway,
                                                                                                               r'''$.ends[0].lon''',
                                                                                                             ),
-                                                                                                            (((getJsonField(
-                                                                                                                              _model.selectedRunway,
-                                                                                                                              r'''$.bearing''',
-                                                                                                                            ) -
-                                                                                                                            180) %
-                                                                                                                        360)
-                                                                                                                    .toInt())
+                                                                                                            ((getJsonField(
+                                                                                                                          _model.selectedRunway,
+                                                                                                                          r'''$.bearing''',
+                                                                                                                        ) -
+                                                                                                                        180) %
+                                                                                                                    360)
                                                                                                                 .toDouble(),
                                                                                                             ((double? elevation) {
-                                                                                                              return elevation != null ? elevation + 1000.0 : null;
+                                                                                                              return elevation != null ? elevation.toDouble() + 1000.0 : null;
                                                                                                             }(getJsonField(
                                                                                                               _model.selectedRunway,
                                                                                                               r'''$.elevation''',
